check if file actually exists before attempting to write data to files dir in wip.

+++ lib/daitss/proc/wip/from_sip.rb
@@ -92,6 +92,8 @@ class Wip
     wip['sip-name'] = sip.name
 
     sip.files.each_with_index do |f, index|
+      next unless File.exists? File.join(sip.path, f)
+
       df = wip.new_original_datafile index
 
       df.open('w') do |o|
